Move files and modules around to make sense in the world of single entry point.
