+++ src/ts/containers/Settings/Settings.tsx
@@ -6,7 +6,7 @@ import {IAppState} from '../../background/store'
 import {Home} from './Home'
 import {FeatureConfig} from './FeatureConfig'
 
-import {Features} from '../../contentScripts/features'
+import {Features} from '../../features/features'
 
 export const Settings = () => {
     const featureId = useSelector<IAppState, string>(state => state.settings.featureId)
