+++ src/ts/features/features.ts
@@ -1,11 +1,11 @@
 import {Feature, prepareSettings, Settings, Shortcut} from '../utils/settings'
 
-import {config as incDec} from './inc-dec-value/index'
-import {config as customCss} from './custom-css/index'
-import {config as srs} from './srs/index'
+import {config as incDec} from './inc-dec-value'
+import {config as customCss} from './custom-css'
+import {config as srs} from '../srs/srs'
 import {config as blockManipulation} from './block-manipulation'
-import {config as estimate} from './estimates/index'
-import {config as navigation} from './navigation/index'
+import {config as estimate} from './estimates'
+import {config as navigation} from './navigation'
 import {filterAsync, mapAsync} from '../utils/async'
 
 export const Features = {
