+++ src/ts/features/fuzzy_date.ts
@@ -1,9 +1,9 @@
 import * as chrono from 'chrono-node'
-import {afterClosingBrackets} from '../../utils/brackets'
-import {RoamDate} from '../../date/common'
-import {RoamNode, Selection} from '../../roam/roam-node'
-import {Roam} from '../../roam/roam'
-import {NodeWithDate} from '../../date/withDate'
+import {afterClosingBrackets} from '../utils/brackets'
+import {RoamDate} from '../date/common'
+import {RoamNode, Selection} from '../roam/roam-node'
+import {Roam} from '../roam/roam'
+import {NodeWithDate} from '../date/withDate'
 
 export const guard = ';'
 const dateContainerExpr = /;(.{3,}?);/gm
