+++ src/ts/srs/srs.ts
@@ -1,9 +1,9 @@
-import {Feature, Shortcut} from '../../utils/settings'
-import {SRSSignal, SRSSignals} from '../../srs/scheduler'
-import {SM2Node} from '../../srs/SM2Node'
-import {AnkiScheduler} from '../../srs/AnkiScheduler'
-import {RoamNode} from '../../roam/roam-node'
-import {Roam} from '../../roam/roam'
+import {Feature, Shortcut} from '../utils/settings'
+import {SRSSignal, SRSSignals} from './scheduler'
+import {SM2Node} from './SM2Node'
+import {AnkiScheduler} from './AnkiScheduler'
+import {RoamNode} from '../roam/roam-node'
+import {Roam} from '../roam/roam'
 
 export const config: Feature = {
     id: 'srs',
